+++ R/flextable.R
@@ -414,6 +414,7 @@ regulartable <- function( data, col_keys = names(data), cwidth = .75, cheight = 
 #' - 'fonts_ignore': if TRUE, pdf-engine 'pdflatex' can be used instead of
 #' 'xelatex' or 'lualatex.' If pdflatex is used, fonts will be ignored because they are
 #' not supported by pdflatex, whereas with the xelatex and lualatex engines they are.
+#' - 'default_line_color': default line color, restored globally after the flextable is produced.
 #' @param word_title alternative text for Word table (used as title of the table)
 #' @param word_description alternative text for Word table (used as description of the table)
 #' @examples
@@ -520,7 +521,9 @@ opts_ft_word <- function(split = get_flextable_defaults()$split, keep_with_next 
 opts_ft_pdf <- function(tabcolsep = get_flextable_defaults()$tabcolsep,
                         arraystretch = get_flextable_defaults()$arraystretch,
                         float = get_flextable_defaults()$float,
-                        fonts_ignore = get_flextable_defaults()$fonts_ignore) {
+                        fonts_ignore = get_flextable_defaults()$fonts_ignore,
+                        default_line_color = "black"
+                        ) {
 
   if( !is.logical(fonts_ignore) || length(fonts_ignore) != 1 ){
     stop(sprintf("'%s' is expected to be a single %s.", "fonts_ignore", "logical"), call. = FALSE)
@@ -538,6 +541,7 @@ opts_ft_pdf <- function(tabcolsep = get_flextable_defaults()$tabcolsep,
   z <- list(tabcolsep = tabcolsep,
             arraystretch = arraystretch,
             float = float,
+            default_line_color = default_line_color,
             fonts_ignore = fonts_ignore)
   class(z) <- "opts_ft_pdf"
   z