+++ tests/testthat/test-borders.R
@@ -1,13 +1,5 @@
 context("check borders rendering")
 
-
-skip_on_cran()
-skip_if_not_installed("doconv")
-library(doconv)
-skip_if_not(doconv::msoffice_available())
-skip_if_not(pandoc_version() >= numeric_version("2"))
-skip_if_not_installed("webshot2")
-
 init_flextable_defaults()
 
 library(data.table)
@@ -51,12 +43,20 @@ ft_1 <- as_flextable(x = tab, separate_with = "VISIT", label_rows = c(LBTEST = "
 ft_1 <- width(ft_1, j = 3, width = 1)
 
 test_that("pptx borders", {
+  skip_if_not(pandoc_version() >= numeric_version("2"))
+  skip_if_not_installed("doconv")
+  library(doconv)
+  skip_if_not(doconv::msoffice_available())
   local_edition(3)
   expect_snapshot_doc(x = save_as_pptx(ft_1, path = tempfile(fileext = ".pptx")),
                       name = "pptx-borders", engine = "testthat")
 })
 
 test_that("docx borders", {
+  skip_if_not(pandoc_version() >= numeric_version("2"))
+  skip_if_not_installed("doconv")
+  library(doconv)
+  skip_if_not(doconv::msoffice_available())
   local_edition(3)
   expect_snapshot_doc(x = save_as_docx(ft_1, path = tempfile(fileext = ".docx")),
                       name = "docx-borders", engine = "testthat")
@@ -64,8 +64,61 @@ test_that("docx borders", {
 
 test_that("html borders", {
   local_edition(3)
+  skip_if_not(pandoc_version() >= numeric_version("2"))
+  skip_if_not_installed("doconv")
+  library(doconv)
+  skip_if_not_installed("webshot2")
   path <- save_as_html(ft_1, path = tempfile(fileext = ".html"))
   expect_snapshot_html(name = "html-borders", path, engine = "testthat")
+})
+
+
+rmd_file_0 <- "rmd/borders.Rmd"
+if (!file.exists(rmd_file_0)) {#just for dev purpose
+  rmd_file_0 <- "tests/testthat/rmd/borders.Rmd"
+}
+rmd_file <- tempfile(fileext = ".Rmd")
+file.copy(rmd_file_0, rmd_file, overwrite = TRUE)
+
+html_file <- gsub("\\.Rmd$", ".html", rmd_file)
+docx_file <- gsub("\\.Rmd$", ".docx", rmd_file)
+pdf_file <- gsub("\\.Rmd$", ".pdf", rmd_file)
+pptx_file <- gsub("\\.Rmd$", ".pptx", rmd_file)
+source("zzzzz.R")
+
+test_that("pdf complex borders", {
+  local_edition(3)
+  library(rmarkdown)
+  skip_if_not(pandoc_available())
+  skip_if_not(pandoc_version() > numeric_version("2.7.3"))
+  render(rmd_file,
+         output_format = rmarkdown::pdf_document(latex_engine = 'xelatex'),
+         output_file = pdf_file,
+         envir = new.env(),
+         quiet = TRUE
+  )
+  expect_snapshot_doc(name = "pdf-complex-borders", pdf_file, engine = "testthat")
+})
+
+test_that("office complex borders", {
+  local_edition(3)
+  library(rmarkdown)
+  skip_if_not(pandoc_available())
+  skip_if_not(pandoc_version() > numeric_version("2.7.3"))
+  render(rmd_file,
+         output_format = rmarkdown::word_document(),
+         output_file = docx_file,
+         envir = new.env(),
+         quiet = TRUE
+  )
+  expect_snapshot_doc(name = "docx-complex-borders", docx_file, engine = "testthat")
+  render(rmd_file,
+         output_format = rmarkdown::powerpoint_presentation(),
+         output_file = pptx_file,
+         envir = new.env(),
+         quiet = TRUE
+  )
+  expect_snapshot_doc(name = "pptx-complex-borders", pptx_file, engine = "testthat")
 })